Add Containerfile, with instructions in comments
This builds a minimal-ish (130MB) container image. It is intended
to be run with two bind mounts: one, read-only, for browser cookies;
the other, writable, the directory into which to create the .puz file.

The Containerfile includes a sample invocation using Firefox. For
use with other browsers, change the .mozilla bind mount and the
first command-line argument as appropriate.

Signed-off-by: Ed Santiago <ed@edsantiago.com>